Possible floor fix

+++ 0x00-python_variable_annotations/2-floor.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+
 """
 2 Basic annotations - floor
 
@@ -6,11 +7,11 @@ a type-annotated function floor which takes a float n as
 argument and returns the floor of the float
 
 Function:
-    floor(n: float) -> float
+    floor(n: float) -> int
 """
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """
     Return the floor of n.
     """
